Seperate client and server parts Needed to split the gui so code compiles correctly on server
